Fixing Bugs and Refactoring Pages Url

+++ pages/jenis-barang/index.php
@@ -25,27 +25,33 @@
         <div class="row">
           <div class="col-12">
             <div class="card">
+              <?php if ($_SESSION['role'] == 'admin') : ?>
               <div class="card-header">
               <div class="float-right">
                   <a href="index.php?page=tambah-jenis-barang" class="btn btn-sm btn-primary"><i class="fas fa-plus"></i>&nbsp; Tambah Jenis Barang</a>
               </div>
               </div>
+              <?php endif; ?>
               <div class="card-body">
                 <table id="example1" class="table table-bordered table-striped">
                   <thead>
                     <th>No</th>
                     <th>Nama Jenis Barang</th>
+                    <?php if ($_SESSION['role'] == 'admin') : ?>
                     <th>Aksi</th>
+                    <?php endif; ?>
                   </thead>
                   <tbody>
                     <?php foreach($result as $jenis):?>
                     <tr>
                       <td><?= $no++ ?></td>
                       <td><?= $jenis['nama_jenis']?></td>
+                      <?php if ($_SESSION['role'] == 'admin') : ?>
                       <td>
                         <a href="index.php?page=edit-jenis-barang&id=<?= $jenis['id_jenis_barang']?>" class="btn btn-sm btn-warning"><i class="fas fa-edit"></i></a>
-                        <a href="index.php?page=hapus-jenis-barang&id=<?= $jenis['id_jenis_barang']?>" class="btn btn-sm btn-danger"><i class="fas fa-trash"></i></a>
+                        <a href="index.php?page=hapus-jenis-barang&id=<?= $jenis['id_jenis_barang']?>" class="btn btn-sm btn-danger" onclick="return confirm('Hapus data?')"><i class="fas fa-trash"></i></a>
                       </td>
+                      <?php endif; ?>
                     </tr>
                     <?php endforeach; ?>
                   </tbody>
